# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "M-Eyes",
    "description": "Open-source DDI platform (DNS, DHCP, IPAM) with Fortinet ecosystem integration",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/networks": {
      "get": {
        "tags": [
          "ipam"
        ],
        "summary": "List Networks",
        "operationId": "list_networks_api_v1_networks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetworkOut"
                  },
                  "type": "array",
                  "title": "Response List Networks Api V1 Networks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "ipam"
        ],
        "summary": "Create Network",
        "operationId": "create_network_api_v1_networks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/networks/{network_id}": {
      "get": {
        "tags": [
          "ipam"
        ],
        "summary": "Get Network",
        "operationId": "get_network_api_v1_networks__network_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Network Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ipam"
        ],
        "summary": "Update Network",
        "operationId": "update_network_api_v1_networks__network_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Network Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ipam"
        ],
        "summary": "Delete Network",
        "operationId": "delete_network_api_v1_networks__network_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Network Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/networks/{network_id}/next-ip": {
      "get": {
        "tags": [
          "ipam"
        ],
        "summary": "Next Ip",
        "operationId": "next_ip_api_v1_networks__network_id__next_ip_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Network Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hosts": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "List Hosts",
        "operationId": "list_hosts_api_v1_hosts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HostOut"
                  },
                  "type": "array",
                  "title": "Response List Hosts Api V1 Hosts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "hosts"
        ],
        "summary": "Create Host",
        "operationId": "create_host_api_v1_hosts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hosts/{host_id}": {
      "delete": {
        "tags": [
          "hosts"
        ],
        "summary": "Delete Host",
        "operationId": "delete_host_api_v1_hosts__host_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "host_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Host Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/system/settings": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get App Settings",
        "operationId": "get_app_settings_api_v1_system_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "system"
        ],
        "summary": "Update App Settings",
        "operationId": "update_app_settings_api_v1_system_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/syslog-test": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Syslog Test",
        "operationId": "syslog_test_api_v1_system_syslog_test_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/system/diagnostics": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Diagnostics",
        "operationId": "diagnostics_api_v1_system_diagnostics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HostIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "network_id": {
            "type": "integer",
            "title": "Network Id"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          },
          "create_reservation": {
            "type": "boolean",
            "title": "Create Reservation",
            "default": false
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "network_id"
        ],
        "title": "HostIn"
      },
      "HostOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ip_address_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address Id"
          },
          "a_record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "A Record Id"
          },
          "ptr_record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ptr Record Id"
          },
          "reservation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservation Id"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "zone_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "ip_address_id",
          "a_record_id",
          "ptr_record_id",
          "reservation_id"
        ],
        "title": "HostOut"
      },
...
      "NetworkIn": {
        "properties": {
          "cidr": {
            "type": "string",
            "title": "Cidr"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "is_container": {
            "type": "boolean",
            "title": "Is Container",
            "default": false
          },
          "vlan": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlan"
          },
          "site": {
            "type": "string",
            "title": "Site",
            "default": ""
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids"
          }
        },
        "type": "object",
        "required": [
          "cidr"
        ],
        "title": "NetworkIn"
      },
      "NetworkOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "cidr": {
            "type": "string",
            "title": "Cidr"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "is_container": {
            "type": "boolean",
            "title": "Is Container"
          },
          "vlan": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlan"
          },
          "site": {
            "type": "string",
            "title": "Site"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "utilization": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UtilizationOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "cidr",
          "name",
          "description",
          "is_container",
          "vlan",
          "site",
          "parent_id"
        ],
        "title": "NetworkOut"
      },
      "NetworkUpdate": {
        "properties": {
          "cidr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cidr"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_container": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Container"
          },
          "vlan": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlan"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids"
          }
        },
        "type": "object",
        "title": "NetworkUpdate"
      },
...
      "SettingsIn": {
        "properties": {
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "values"
        ],
        "title": "SettingsIn"
      },
      "SettingsOut": {
        "properties": {
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "values"
        ],
        "title": "SettingsOut"
      },
...
      "TagOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color"
        ],
        "title": "TagOut"
      },
...
      "UtilizationOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "dhcp_range": {
            "type": "integer",
            "title": "Dhcp Range"
          },
          "free": {
            "type": "integer",
            "title": "Free"
          },
          "percent": {
            "type": "number",
            "title": "Percent"
          }
        },
        "type": "object",
        "required": [
          "total",
          "used",
          "dhcp_range",
          "free",
          "percent"
        ],
        "title": "UtilizationOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}